first_class/requirements.txt
asgiref==3.8.1
Django==5.0.4
sqlparse==0.4.4
tzdata==2024.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-phonenumber-field[phonenumbers]